requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
discord.py>=2.7.1
python-dotenv>=1.2.2
pydantic-settings>=2.6.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.2.0
celery>=5.4.0
redis>=5.0.0
